requirements.txt
streamlit
deepgram-sdk
openai
transformers
torch
streamlit-mic-recorder
